hub/requirements.txt
asyncio
python-socketio
async-timeout
aiohttp
fastapi
boto3